Updated edgeDetection.m
Created find_edges function, gradient_magnitude function, and spatial_filter function.

Each function returns a value and they are further explained in the project report.
The program takes user input to set a scalar threshold value, which is sent to the find_edges function to detect edges specified by the user. An image is shown to the user based on their inputs and the program will continue to prompt the user for another scalar value. A value of '-1' quits the program. The program exits with a message telling the user that the program is quitting.

The program takes one image input. It can continue printing images as long as the user continues using the program.